#32 管理画面一部リファクタリング
